More progress on parser. Simplified keywords
Keywords should be simplified even more though as they are the most basic type in the entire parser. We don't need very much information about the word itself. Or we need to wrap the keyword type in a way that we can easily do evaluations on it and set expectations.